Extract out some types from real traffic stream module
